refactor(THU-808): fold publish-then-persist into one language setter
